[ISSUE-2] Add page to display who is living
This still has some structures I don't like but it works.
No tests yet.
